feat: Improve reserve logs (#505)
