+++ src/pages/pipeline-details/components/ReserveLogs/components/AllLogs/types.ts
@@ -1,0 +1,5 @@
+import { NormalizedLogs } from '../../../../providers/DynamicData/types';
+
+export interface AllLogsProps {
+  logs: NormalizedLogs;
+}
