+++ src/pages/pipeline-details/components/ReserveLogs/components/LogsByTask/types.ts
@@ -1,0 +1,5 @@
+import { NormalizedLogs } from '../../../../providers/DynamicData/types';
+
+export interface LogsByTaskProps {
+  logs: NormalizedLogs;
+}
