--- src/pages/pipeline-details/components/ReserveLogs/types.ts
@@ -1,5 +1,0 @@
-import { NormalizedLogs } from '../../providers/DynamicData/types';
-
-export interface ReserveLogsProps {
-  logs: NormalizedLogs;
-}
